Fix type data of annotation custom

+++ src/main/java/com/example/haus/domain/validator/PositiveOrNullValidator.java
@@ -3,9 +3,9 @@ package com.example.haus.domain.validator;
 import jakarta.validation.ConstraintValidator;
 import jakarta.validation.ConstraintValidatorContext;
 
-public class PositiveOrNullValidator implements ConstraintValidator<PositiveOrNull, Float> {
+public class PositiveOrNullValidator implements ConstraintValidator<PositiveOrNull, Long> {
     @Override
-    public boolean isValid(Float value, ConstraintValidatorContext context) {
+    public boolean isValid(Long value, ConstraintValidatorContext context) {
         return value == null || value >= 0;
     }
